#1: Refactor packages
