+++ ui/src/main/java/com/octogonapus/omj/ui/model/MethodTrace.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with OMJ.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.octogonapus.omj.ui;
+package com.octogonapus.omj.ui.model;
 
 import java.util.List;
